+++ core/lib/Thelia/Model/Lang.php
@@ -5,11 +5,8 @@ namespace Thelia\Model;
 use Propel\Runtime\Connection\ConnectionInterface;
 use Propel\Runtime\Exception\PropelException;
 use Propel\Runtime\Propel;
-use Symfony\Component\Filesystem\Filesystem;
 use Thelia\Core\Event\Lang\LangEvent;
 use Thelia\Core\Event\TheliaEvents;
-use Thelia\Core\Template\TemplateHelper;
-use Thelia\Core\Translation\Translator;
 use Thelia\Model\Base\Lang as BaseLang;
 use Thelia\Model\Map\LangTableMap;
 
@@ -64,34 +61,6 @@ class Lang extends BaseLang
         }
     }
 
-    protected function fixMissingFlag()
-    {
-        // Be sure that a lang have a flag, otherwise copy the
-        // "unknown" flag
-        $adminTemplate = TemplateHelper::getInstance()->getActiveAdminTemplate();
-        $unknownFlag = ConfigQuery::getUnknownFlagPath();
-
-        $unknownFlagPath = $adminTemplate->getAbsolutePath().DS.$unknownFlag;
-
-        if (! file_exists($unknownFlagPath)) {
-            throw new \RuntimeException(
-                Translator::getInstance()->trans(
-                    "The image which replaces an undefined country flag (%file) was not found. Please check unknown-flag-path configuration variable, and check that the image exists.",
-                    array("%file" => $unknownFlag)
-                )
-            );
-        }
-
-        // Check if the country flag exists
-        $countryFlag = rtrim(dirname($unknownFlagPath), DS).DS.$this->getCode().'.png';
-
-        if (! file_exists($countryFlag)) {
-            $fs = new Filesystem();
-
-            $fs->copy($unknownFlagPath, $countryFlag);
-        }
-    }
-
     public function preInsert(ConnectionInterface $con = null)
     {
         $this->dispatchEvent(TheliaEvents::BEFORE_CREATELANG, new LangEvent($this));
@@ -103,7 +72,7 @@ class Lang extends BaseLang
     {
         $this->dispatchEvent(TheliaEvents::AFTER_CREATELANG, new LangEvent($this));
 
-        $this->fixMissingFlag();
+        $this->dispatchEvent(TheliaEvents::LANG_FIX_MISSING_FLAG, new LangEvent($this));
     }
 
     public function preUpdate(ConnectionInterface $con = null)
@@ -117,7 +86,7 @@ class Lang extends BaseLang
     {
         $this->dispatchEvent(TheliaEvents::AFTER_UPDATELANG, new LangEvent($this));
 
-        $this->fixMissingFlag();
+        $this->dispatchEvent(TheliaEvents::LANG_FIX_MISSING_FLAG, new LangEvent($this));
     }
 
     public function preDelete(ConnectionInterface $con = null)