+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 - new 'asset' Smarty function, to get the URL of an arbitrary file from template assets, such as a video or a font.
 - Imagine package is updated to 0.6.2, which provides a better support for transparency.
 - Default border color of images resized with resize_mode="border" is now transparent instead of opaque white.
+- The TemplateHelper class is deprecated. You should now use the thelia.template_helper service. TemplateHelperInterface has been introduced, so that modules may implement alternate versions
+
 
 # 2.1.3